+++ src/main/java/de/tum/in/www1/artemis/web/rest/ProgrammingExerciseResource.java
@@ -597,7 +597,8 @@ public class ProgrammingExerciseResource {
         }
 
         // Forbid changes of course exercise belongs to.
-        if (programmingExerciseBeforeUpdate.getCourseViaExerciseGroupOrCourseMember().getId() != updatedProgrammingExercise.getCourseViaExerciseGroupOrCourseMember().getId()) {
+        if (programmingExerciseBeforeUpdate.getCourseViaExerciseGroupOrCourseMember().getId().longValue() != updatedProgrammingExercise.getCourseViaExerciseGroupOrCourseMember()
+                .getId().longValue()) {
             return conflict("The programming exercise course cannot be changed", ENTITY_NAME, "cannotChangeCourseId");
         }
